initialise go module
radau is incompatible with current versions of go-pg/pg and
go-pg/migrations. To support current builds, this commit introduces a
go.mod file and pinpoints both dependencies to older versions (which
were current around the last time of a radau release). This re-enables
builds on current versions of Golang, but in the long run, radau should
be updated to work with the latest versions of its dependencies.